# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ChillBooking API",
        "description": "API documentation for ChillBooking service",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/location-schedules": {
            "get": {
                "tags": [
                    "Location Schedules"
                ],
                "summary": "Retrieve list of location schedules",
                "operationId": "c4500500969a68cd8b3c0bc92efbe1b6",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LocationScheduleResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Location Schedules"
                ],
                "summary": "Create a new location schedule",
                "operationId": "88e9d91671580a5281dd5bf88ed8c256",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LocationScheduleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Location schedule created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LocationScheduleResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/location-schedules/{id}": {
            "get": {
                "tags": [
                    "Location Schedules"
                ],
                "summary": "Retrieve a specific location schedule",
                "operationId": "d4c613644853ec5a3f44813f2296feff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LocationScheduleResource"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Location Schedules"
                ],
                "summary": "Update a specific location schedule",
                "operationId": "4f39fbc936a6fb2065706ed6d32ab3b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LocationScheduleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Location schedule updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LocationScheduleResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Location Schedules"
                ],
                "summary": "Delete a specific location schedule",
                "operationId": "2534ed21795f6b05a8822f02c0ecc8b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Location schedule deleted successfully"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "LocationScheduleRequest": {
                "required": [
                    "location_id",
                    "day_number"
                ],
                "properties": {
                    "location_id": {
                        "description": "ID of the location",
                        "type": "integer"
                    },
                    "day_number": {
                        "description": "Day number (0-6)",
                        "type": "integer",
                        "maximum": 6,
                        "minimum": 0
                    },
                    "from": {
                        "description": "Opening time",
                        "type": "string",
                        "format": "time",
                        "nullable": true
                    },
                    "to": {
                        "description": "Closing time",
                        "type": "string",
                        "format": "time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "LocationResource": {
                "properties": {
                    "id": {
                        "description": "ID of the location",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Title of the location",
                        "type": "string"
                    },
                    "lng": {
                        "description": "Longitude of the location",
                        "type": "number",
                        "format": "double",
                        "nullable": true
                    },
                    "lat": {
                        "description": "Latitude of the location",
                        "type": "number",
                        "format": "double",
                        "nullable": true
                    },
                    "address": {
                        "description": "Address of the location",
                        "type": "string",
                        "nullable": true
                    },
                    "city": {
                        "description": "City of the location",
                        "type": "string",
                        "nullable": true
                    },
                    "postcode": {
                        "description": "Postcode of the location",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "schedules": {
                        "description": "List of schedules for the location",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LocationScheduleResource"
                        }
                    }
                },
                "type": "object"
            },
            "LocationScheduleResource": {
                "properties": {
                    "id": {
                        "description": "ID of the schedule",
                        "type": "integer"
                    },
                    "location_id": {
                        "description": "ID of the location",
                        "type": "integer"
                    },
                    "day_number": {
                        "description": "Day number (0-6)",
                        "type": "integer"
                    },
                    "from": {
                        "description": "Opening time",
                        "type": "string",
                        "format": "time",
                        "nullable": true
                    },
                    "to": {
                        "description": "Closing time",
                        "type": "string",
                        "format": "time",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "location": {
                        "$ref": "#/components/schemas/LocationResource"
                    }
                },
                "type": "object"
            }
        }
    },
...
}